# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: api
  version: 1.0.0
  description: Shawrome Car Sales Platform API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CarStatus:
      type: string
      enum: [available, reserved, sold]
    CarFuel:
      type: string
      enum: [essence, diesel, hybride, electrique, gpl]
    CarTransmission:
      type: string
      enum: [manuelle, automatique]
    CarCondition:
      type: string
      enum: [neuf, occasion]
    Car:
      type: object
      required: [id, brand, model, year, mileage, purchasePrice, sellingPrice, fuel, transmission, color, condition, status, photos, sellerId, sellerName, createdAt]
      properties:
        id:
          type: integer
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        description:
          type: string
          nullable: true
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        fuel:
          $ref: '#/components/schemas/CarFuel'
        transmission:
          $ref: '#/components/schemas/CarTransmission'
        color:
          type: string
        condition:
          $ref: '#/components/schemas/CarCondition'
        status:
          $ref: '#/components/schemas/CarStatus'
        photos:
          type: array
          items:
            type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        soldAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateCarRequest:
      type: object
      required: [brand, model, year, mileage, purchasePrice, sellingPrice, fuel, transmission, color, condition]
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        description:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        fuel:
          $ref: '#/components/schemas/CarFuel'
        transmission:
          $ref: '#/components/schemas/CarTransmission'
        color:
          type: string
        condition:
          $ref: '#/components/schemas/CarCondition'
        sellerId:
          type: integer
    UpdateCarRequest:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: integer
        description:
          type: string
        purchasePrice:
          type: number
        sellingPrice:
          type: number
        fuel:
          $ref: '#/components/schemas/CarFuel'
        transmission:
          $ref: '#/components/schemas/CarTransmission'
        color:
          type: string
        condition:
          $ref: '#/components/schemas/CarCondition'
        sellerId:
          type: integer
    UpdateCarStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/CarStatus'
    UpdateCarPriceRequest:
      type: object
      required: [sellingPrice]
      properties:
        sellingPrice:
          type: number
    UploadPhotosResponse:
      type: object
      required: [photos]
      properties:
        photos:
          type: array
          items:
            type: string
    CarsListResponse:
      type: object
      required: [cars, total, page, limit]
      properties:
        cars:
          type: array
          items:
            $ref: '#/components/schemas/Car'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ──────────────────────────── Price History ────────────────────────────
...
    Sale:
      type: object
      required: [id, carId, carBrand, carModel, carYear, finalPrice, purchasePrice, profit, profitPercent, sellerId, sellerName, soldAt]
      properties:
        id:
          type: integer
        carId:
          type: integer
        carBrand:
          type: string
        carModel:
          type: string
        carYear:
          type: integer
        finalPrice:
          type: number
        purchasePrice:
          type: number
        profit:
          type: number
        profitPercent:
          type: number
        sellerId:
          type: integer
        sellerName:
          type: string
        soldAt:
          type: string
          format: date-time
...
    UploadCarPhotosRequest:
      type: object
      required: [photos]
      properties:
        photos:
          type: array
          items:
            type: string
          description: Base64 encoded image strings
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
paths:
...
  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List cars with filters and pagination
      parameters:
        - name: brand
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CarStatus'
        - name: fuel
          in: query
          schema:
            $ref: '#/components/schemas/CarFuel'
        - name: transmission
          in: query
          schema:
            $ref: '#/components/schemas/CarTransmission'
        - name: condition
          in: query
          schema:
            $ref: '#/components/schemas/CarCondition'
        - name: yearMin
          in: query
          schema:
            type: integer
        - name: yearMax
          in: query
          schema:
            type: integer
        - name: priceMin
          in: query
          schema:
            type: number
        - name: priceMax
          in: query
          schema:
            type: number
        - name: mileageMax
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of cars
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarsListResponse'

    post:
      operationId: createCar
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Create a new car listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCarRequest'
      responses:
        '201':
          description: Car created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Car detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: updateCar
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Update car (admin or assigned seller)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCarRequest'
      responses:
        '200':
          description: Car updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'

    delete:
      operationId: deleteCar
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Delete car (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Car deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /cars/{id}/status:
    patch:
      operationId: updateCarStatus
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Update car status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCarStatusRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
  /cars/{id}/price:
    patch:
      operationId: updateCarPrice
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Update car selling price (logs to price history)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCarPriceRequest'
      responses:
        '200':
          description: Price updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
  /cars/{id}/photos:
    post:
      operationId: uploadCarPhotos
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Upload photos for a car (sends base64 encoded images)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadCarPhotosRequest'
      responses:
        '200':
          description: Photos uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadPhotosResponse'
  /cars/{id}/sell:
    post:
      operationId: markCarAsSold
      tags: [cars]
      security:
        - bearerAuth: []
      summary: Mark a car as sold (creates a sale record and computes profit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Car marked as sold, sale record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'

  # ──────────────────────────── Price History ────────────────────────────
...