# lib/api-zod/src/generated/types/carCondition.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Shawrome Car Sales Platform API
 * OpenAPI spec version: 1.0.0
 */

export type CarCondition = typeof CarCondition[keyof typeof CarCondition];


export const CarCondition = {
  neuf: 'neuf',
  occasion: 'occasion',
} as const;